requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
audioread==3.1.0
babel==2.18.0
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
comm==0.2.3
confection==1.3.3
config==0.5.1
cryptography==46.0.7
csvw==3.7.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
datasets==4.8.4
dateparser==1.4.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
executing==2.2.1
fastapi==0.135.2
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
google-auth==2.49.2
google-genai==1.73.1
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
HyperPyYAML==1.2.3
idna==3.11
IProgress==0.4
ipykernel==6.31.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kokoro==0.9.4
language-tags==1.2.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.20.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.24.4
onnxruntime-gpu==1.24.4
openai-whisper==20250625
openwakeword==0.4.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
phonemizer==3.3.0
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.9.4
pooch==1.9.0
preshed==3.0.13
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
pyxdg==0.28
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.15.7
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
segments==2.4.0
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.13
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.1.0
spyder-kernels==3.1.3
srsly==2.5.3
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
websockets==15.0.1
widgetsnbextension==4.0.15
wrapt==2.1.2
wurlitzer==3.1.1
xxhash==3.6.0
yarl==1.23.0
